refactor: move authentication and data helper functions to a new lib directory; remove unused utils files
