+++ server/lib/auth.helper.ts
